requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.7.4
click==8.1.7
Distance==0.1.3
distro==1.9.0
filelock==3.15.4
fsspec==2024.6.1
g2p-en==2.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
inflect==7.3.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
MarkupSafe==2.1.5
more-itertools==10.4.0
mpmath==1.3.0
networkx==3.3
nltk==3.9.1
numpy==2.1.0
openai==1.42.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
python-dotenv==1.0.1
regex==2024.7.24
sniffio==1.3.1
sympy==1.13.2
torch==2.2.1
torchaudio==2.2.1
tqdm==4.66.5
typeguard==4.3.0
typing_extensions==4.12.2
